Tidy up dependencies (#41)
* Tidy up dependencies

* scalafmt
