+++ project/plugins.sbt
@@ -1,6 +1,7 @@
-addSbtPlugin("com.dwijnand"       % "sbt-dynver"        % "4.0.0")
-addSbtPlugin("org.scalameta"      % "sbt-scalafmt"      % "2.0.2")
-addSbtPlugin("pl.project13.scala" % "sbt-jmh"           % "0.3.7")
-addSbtPlugin("com.47deg"          % "sbt-microsites"    % "0.9.2")
-addSbtPlugin("ch.epfl.scala"      % "sbt-release-early" % "2.1.1")
-addSbtPlugin("org.scoverage"      % "sbt-scoverage"     % "1.6.0")
+addSbtPlugin("com.dwijnand"              % "sbt-dynver"        % "4.0.0")
+addSbtPlugin("org.scalameta"             % "sbt-scalafmt"      % "2.0.2")
+addSbtPlugin("pl.project13.scala"        % "sbt-jmh"           % "0.3.7")
+addSbtPlugin("com.47deg"                 % "sbt-microsites"    % "0.9.2")
+addSbtPlugin("ch.epfl.scala"             % "sbt-release-early" % "2.1.1")
+addSbtPlugin("org.scoverage"             % "sbt-scoverage"     % "1.6.0")
+addSbtPlugin("io.github.davidgregory084" % "sbt-tpolecat"      % "0.1.7")
